chore(react-menu): extract useMenuItemBase_unstable into a separate module
