+++ change/@fluentui-react-menu-f088e693-cbb1-405b-b804-a0c5ca5dc15a.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "chore: extract useMenuItemBase_unstable into a separate module",
+  "packageName": "@fluentui/react-menu",
+  "email": "viktorgenaev@microsoft.com",
+  "dependentChangeType": "patch"
+}
