+++ packages/react-components/react-menu/library/src/components/MenuItem/index.ts
@@ -1,5 +1,6 @@
 export { MenuItem } from './MenuItem';
 export type { MenuItemProps, MenuItemSlots, MenuItemState } from './MenuItem.types';
 export { renderMenuItem_unstable } from './renderMenuItem';
-export { useMenuItem_unstable, useMenuItemBase_unstable } from './useMenuItem';
+export { useMenuItem_unstable } from './useMenuItem';
+export { useMenuItemBase_unstable } from './useMenuItemBase';
 export { menuItemClassNames, useMenuItemStyles_unstable } from './useMenuItemStyles.styles';
